Moved utils to JMESPath specific file
